Don't use sqrt for rotors; use sqrtOfRotor
When constructing a frame from knowledge of the Z axis (FrameFromZ),
it is possible that Z=-z, in which case the sqrt algorithm divides by
zero.  This is because there is no unique shortest rotation taking z
to -z.  Instead, it is more appropriate to arbitrarily choose one such
rotation, and use that as the sqrt.  This is what sqrtOfRotor does.